Fix indefinite execution of the same project
Uploading a empty file caused the project to be run indefinitely.
As a result, other projects in the queue were on hold for an
indefinite amount of time. This patch fixed this by checking
the `data` section of the config (whether it is empty) and halt
the project execution at once.